update read me with todo's
