+++ README.md
@@ -1,5 +1,10 @@
-# README
+### Stuff to do and Notes
 
-git remote add origin git@github.com:jaredsurya/regulate-app.git
-git branch -M main
-git push -u origin main
+- Route interal pages somehow, learn this
+  - React Navigation
+- ERM for backend, users, journal, meditation
+- Create login or home page for styling practice
+- complete some wireframing
+- Styling, addition of libraries or CSS
+- Route backend and hit API's
+  - Auth pages ready for hitting backend
